# Switch to sync.Map for cached event stats