@axw 2023-07-20
Benchmarks are from before I fixed releasing back to the pool, but it should not make a difference.