@carsonip 2023-07-20
Benchmarks from my laptop with the 2nd commit:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after │
                            │                sec/op                 │    sec/op      vs base               │
AggregateCombinedMetrics-16                            3.293µ ±  3%    3.475µ ± 17%  +5.53% (p=0.000 n=10)
AggregateBatchSerial-16                                15.62µ ±  5%    15.92µ ±  4%  +1.94% (p=0.019 n=10)
AggregateBatchParallel-16                              16.79µ ±  3%    17.04µ ±  3%       ~ (p=0.739 n=10)
CombinedMetricsEncoding-16                             4.983µ ±  1%    5.104µ ±  3%  +2.43% (p=0.007 n=10)
CombinedMetricsDecoding-16                             4.412µ ±  1%    4.645µ ±  1%  +5.28% (p=0.000 n=10)
CombinedMetricsToBatch-16                              418.5µ ±  3%    416.6µ ±  2%       ~ (p=0.631 n=10)
EventToCombinedMetrics-16                              1.992µ ± 26%    2.097µ ± 12%       ~ (p=0.315 n=10)
geomean                                                10.68µ          11.00µ        +3.05%

                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after  │
                            │                 B/op                  │     B/op      vs base                 │
AggregateCombinedMetrics-16                          3.634Ki ± 3%     3.702Ki ± 3%  +1.87% (p=0.019 n=10)
AggregateBatchSerial-16                              23.83Ki ± 0%     24.03Ki ± 1%  +0.83% (p=0.000 n=10)
AggregateBatchParallel-16                            23.87Ki ± 0%     23.98Ki ± 0%  +0.45% (p=0.000 n=10)
CombinedMetricsEncoding-16                             0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                           10.01Ki ± 0%     10.01Ki ± 0%       ~ (p=1.000 n=10)
CombinedMetricsToBatch-16                            39.77Ki ± 0%     37.66Ki ± 0%  -5.30% (p=0.000 n=10)
EventToCombinedMetrics-16                            4.188Ki ± 0%     4.188Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                           ²                 -0.33%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after │
                            │               allocs/op               │  allocs/op   vs base                 │
AggregateCombinedMetrics-16                            27.00 ± 0%      29.00 ± 3%  +7.41% (p=0.000 n=10)
AggregateBatchSerial-16                                106.0 ± 0%      112.0 ± 0%  +5.66% (p=0.000 n=10)
AggregateBatchParallel-16                              106.0 ± 0%      112.0 ± 0%  +5.66% (p=0.000 n=10)
CombinedMetricsEncoding-16                             0.000 ± 0%      0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                             40.00 ± 0%      40.00 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsToBatch-16                              339.0 ± 0%      308.0 ± 0%  -9.14% (p=0.000 n=10)
EventToCombinedMetrics-16                              17.00 ± 0%      17.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                           ²                +1.23%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```