@carsonip 2023-07-20
Benchmark:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after │
                            │                sec/op                 │    sec/op      vs base               │
AggregateCombinedMetrics-16                             3.350µ ± 3%    3.352µ ± 19%       ~ (p=0.841 n=10)
AggregateBatchSerial-16                                 15.68µ ± 3%    16.04µ ±  3%       ~ (p=0.190 n=10)
AggregateBatchParallel-16                               16.72µ ± 4%    17.24µ ±  3%       ~ (p=0.105 n=10)
CombinedMetricsEncoding-16                              5.066µ ± 3%    5.021µ ±  2%  -0.91% (p=0.035 n=10)
CombinedMetricsDecoding-16                              4.505µ ± 0%    4.500µ ±  2%       ~ (p=0.542 n=10)
CombinedMetricsToBatch-16                               409.2µ ± 2%    407.8µ ±  2%       ~ (p=0.853 n=10)
EventToCombinedMetrics-16                               1.977µ ± 1%    1.979µ ±  1%       ~ (p=0.811 n=10)
geomean                                                 10.71µ         10.78µ        +0.59%

                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after  │
                            │                 B/op                  │     B/op      vs base                 │
AggregateCombinedMetrics-16                          3.709Ki ± 1%     3.707Ki ± 0%       ~ (p=1.000 n=10)
AggregateBatchSerial-16                              24.01Ki ± 1%     23.93Ki ± 1%       ~ (p=0.315 n=10)
AggregateBatchParallel-16                            23.97Ki ± 0%     23.97Ki ± 0%       ~ (p=0.912 n=10)
CombinedMetricsEncoding-16                             0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                           10.01Ki ± 0%     10.01Ki ± 0%       ~ (p=0.628 n=10)
CombinedMetricsToBatch-16                            37.66Ki ± 0%     37.66Ki ± 0%       ~ (p=1.000 n=10)
EventToCombinedMetrics-16                            4.188Ki ± 0%     4.188Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                           ²                 -0.06%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ bench-out/syncmap-cached-stats-before │ bench-out/syncmap-cached-stats-after │
                            │               allocs/op               │  allocs/op   vs base                 │
AggregateCombinedMetrics-16                            29.00 ± 0%      29.00 ± 0%       ~ (p=1.000 n=10) ¹
AggregateBatchSerial-16                                112.0 ± 0%      112.0 ± 0%       ~ (p=1.000 n=10) ¹
AggregateBatchParallel-16                              112.0 ± 0%      112.0 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsEncoding-16                             0.000 ± 0%      0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                             40.00 ± 0%      40.00 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsToBatch-16                              308.0 ± 0%      308.0 ± 0%       ~ (p=1.000 n=10)
EventToCombinedMetrics-16                              17.00 ± 0%      17.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                           ²                +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```